+++ mod.ts
@@ -1,6 +1,7 @@
 export * from "./runner.ts";
 export * from "./runners/proc-group.ts";
 export * from "./runners/handlers/string.ts";
+export * from "./runners/handlers/string-array.ts";
 export * from "./runners/handlers/bytes.ts";
 export * from "./runners/handlers/reader.ts";
 export * from "./runners/handlers/string-iterable.ts";
@@ -9,5 +10,6 @@ export * from "./runners/handlers/empty.ts";
 export * from "./runners/handlers/stderr-to-stdout-string-iterable.ts";
 export * from "./runners/closers.ts";
 export * from "./runners/process-exit-error.ts";
+export * from "./runners/chained-error.ts";
 export * from "./runners/stderr-support.ts";
 export { sleep } from "./runners/utility.ts";