+++ runners/handlers/string-array-handlers.ts
@@ -1,28 +1,30 @@
-import { StringReader } from "../../deps.ts";
 import { ChainedError } from "../chained-error.ts";
 import {
   MultiCloseProcess,
   MultiCloseReader,
   MultiCloseWriter,
-  NoCloseReader,
 } from "../closers.ts";
 import { ErrorHandler } from "../error-support.ts";
 import { InputHandler } from "../proc-group.ts";
 import { StderrProcessor } from "../stderr-support.ts";
-import { pump } from "../utility.ts";
 import { AbstractTextOutputHandler } from "./abstract-handlers.ts";
 
 /**
- * Source `stdin` from a `string`. `stdin` is closed once the text data is written.
+ * Source `stdin` from a `string[]`. `stdin` is closed once the text data is written.
  */
-export class StringInputHandler implements InputHandler<string> {
+export class StringArrayInputHandler implements InputHandler<string[]> {
   get failOnEmptyInput(): boolean {
     return true;
   }
 
-  async processInput(input: string, stdin: MultiCloseWriter): Promise<void> {
+  async processInput(input: string[], stdin: MultiCloseWriter): Promise<void> {
     try {
-      await pump(new NoCloseReader(new StringReader(input)), stdin);
+      const encoder = new TextEncoder();
+      const lf = encoder.encode("\n");
+      for (const line of input) {
+        await stdin.write(encoder.encode(line));
+        await stdin.write(lf);
+      }
     } catch (e) {
       throw new ChainedError(`${this.constructor.name}.processInput`, e);
     } finally {
@@ -32,9 +34,10 @@ export class StringInputHandler implements InputHandler<string> {
 }
 
 /**
- * Return `stdout` as a `string`.
+ * Return `stdout` as a `string[]`.
  */
-export class StringOutputHandler extends AbstractTextOutputHandler<string> {
+export class StringArrayOutputHandler
+  extends AbstractTextOutputHandler<string[]> {
   constructor(
     processStderr: StderrProcessor,
     errorHandler: ErrorHandler,
@@ -47,13 +50,13 @@ export class StringOutputHandler extends AbstractTextOutputHandler<string> {
     stderr: MultiCloseReader,
     process: MultiCloseProcess,
     input: { stdin: MultiCloseWriter; handlerResult: Promise<null | Error> },
-  ): Promise<string> {
+  ): Promise<string[]> {
     const lines = [];
 
     for await (const line of this.process(stdout, stderr, process, input)) {
       lines.push(line);
     }
 
-    return lines.join("\n");
+    return lines;
   }
 }