+++ runners/process-exit-error.ts
@@ -15,9 +15,9 @@ export class ProcessExitError extends Error {
     /** Signal, if the exit was due to a signal. */
     public readonly signal?: number,
     /** Details (optionally scraped from `stderr`). */
-    public readonly details?: string,
+    public readonly details?: string[],
   ) {
-    super(message);
+    super(`${message} ${JSON.stringify(options.cmd)}`);
     this.name = this.constructor.name;
   }
 }